+++ client/.env
@@ -1,0 +1,1 @@
+REACT_APP_SERVER_HOST=https://103f-121-187-22-182.ngrok-free.app
