+++ client/src/api/useFetch.ts
@@ -1,8 +1,10 @@
 import { PerformanceListType, PerformanceType } from '../model/Performance';
-import { ArtistList } from '../model/Artist';
+import { ArtistList, Artist } from '../model/Artist';
+import { Member, Performance, Review } from '../model/Member';
 import { useEffect, useState } from 'react';
 import axios from 'axios';
 import { useNavigate } from 'react-router-dom';
+import { getCookie } from '../utils/Cookie';
 
 const TEST_HOST = process.env.REACT_APP_TEST_HOST;
 const SERVER_HOST = process.env.REACT_APP_SERVER_HOST;
@@ -76,7 +78,7 @@ export const useGetPerformance = (id: string | number | undefined) => {
 
   const getData = async () => {
     await axios
-      .get<PerformanceType>(`${TEST_HOST}/performances/${id}`, {
+      .get<PerformanceType>(`${SERVER_HOST}/performance/${id}`, {
         headers: { 'ngrok-skip-browser-warning': true },
       })
       .then(data => setData(data.data))
@@ -95,6 +97,83 @@ export const useGetArtists = () => {
   const getData = async () => {
     await axios
       .get<ArtistList>(`${SERVER_HOST}/artist?category=1&page=1&size=10`, {
+        headers: { 'ngrok-skip-browser-warning': true },
+      })
+      .then(data => setData(data.data))
+      .catch(err => console.log(err));
+  };
+  useEffect(() => {
+    getData();
+  }, []);
+
+  return data;
+};
+
+export const useGetArtist = (id: string | number | undefined) => {
+  const [data, setData] = useState<Artist>();
+
+  const getData = async () => {
+    await axios
+      .get<Artist>(`${SERVER_HOST}/artist/${id}`, {
+        headers: { 'ngrok-skip-browser-warning': true },
+      })
+      .then(data => setData(data.data))
+      .catch(err => console.log(err));
+  };
+  useEffect(() => {
+    getData();
+  }, []);
+
+  return data;
+};
+
+export const useGetMember = () => {
+  const [data, setData] = useState<Member>();
+
+  const getData = async () => {
+    await axios
+      .get<Member>(`${SERVER_HOST}/member`, {
+        headers: {
+          Authorization: getCookie('accessToken'),
+          'ngrok-skip-browser-warning': true,
+        },
+      })
+      .then(data => setData(data.data))
+      .catch(err => console.log(err));
+  };
+  useEffect(() => {
+    getData();
+  }, []);
+
+  return data;
+};
+
+export const useGetMemberPerformance = (id: string | number | undefined) => {
+  const [data, setData] = useState<Performance>();
+
+  const getData = async () => {
+    await axios
+      // 공연받아오는 endpoint에 맞게 수정해주기
+      .get<Performance>(`${SERVER_HOST}/member/${id}`, {
+        headers: { 'ngrok-skip-browser-warning': true },
+      })
+      .then(data => setData(data.data))
+      .catch(err => console.log(err));
+  };
+  useEffect(() => {
+    getData();
+  }, []);
+
+  return data;
+};
+
+export const useGetMemberReview = (id: string | number | undefined) => {
+  const [data, setData] = useState<Review>();
+
+  const getData = async () => {
+    await axios
+      // 공연받아오는 endpoint에 맞게 수정해주기
+      .get<Review>(`${SERVER_HOST}/member/${id}`, {
         headers: { 'ngrok-skip-browser-warning': true },
       })
       .then(data => setData(data.data))
