+++ client/src/components/header/Header.tsx
@@ -3,6 +3,10 @@ import TicketIcon from '../../icons/TicketIcon';
 import { useNavigate } from 'react-router-dom';
 import PreviousIcon from '../../icons/PreviousIcon';
 import { Styled_Header } from './Header.styled';
+import NoTicketModal from '../modal/NoTicket';
+import TicketModal from '../modal/Ticket';
+import { isStyledComponent } from 'styled-components';
+import { useState } from 'react';
 
 interface OwnProps {
   precious?: boolean;
@@ -10,6 +14,21 @@ interface OwnProps {
 
 const Header = ({ precious }: OwnProps) => {
   const navigate = useNavigate();
+  const [isOpen, setIsOpen] = useState(false);
+
+  const handleModalOpen = () => {
+    setIsOpen(true);
+  };
+  const handleModalClose = () => {
+    setIsOpen(false);
+  };
+  const handleModalClick = (
+    e: React.MouseEvent<HTMLDivElement, MouseEvent>
+  ) => {
+    if (e.target === e.currentTarget) {
+      handleModalClose();
+    }
+  };
 
   return (
     <>
@@ -30,7 +49,7 @@ const Header = ({ precious }: OwnProps) => {
                 <Styled_Header.SearchS>
                   <SearchIcon />
                 </Styled_Header.SearchS>
-                <Styled_Header.TicketS>
+                <Styled_Header.TicketS onClick={handleModalOpen}>
                   <TicketIcon />
                 </Styled_Header.TicketS>
               </Styled_Header.Div>
@@ -38,6 +57,8 @@ const Header = ({ precious }: OwnProps) => {
           )}
         </Styled_Header.Nav>
       </Styled_Header.Header>
+      {/* 멤버가 예약한 공연이 있으면 */}
+      {isOpen && <NoTicketModal onClick={handleModalClick} />}
     </>
   );
 };
