+++ client/src/components/modal/Reservation.tsx
@@ -45,7 +45,7 @@ const S = {
     left: 0;
     width: 100%;
     height: 100%;
-    /* background-color: rgba(0, 0, 0, 0.5); */
+    z-index: 1;
     display: flex;
     justify-content: center;
     align-items: center;
