+++ client/src/pages/performance-info/PerformanceInfo.style.ts
@@ -28,7 +28,7 @@ export const Container = styled.div`
 `;
 export const Main = styled.main`
   width: 390px;
-  min-height: calc(100vh - 50px);
+  min-height: calc(100vh - 60px);
   background-color: var(--theme-background-color);
   display: flex;
   flex-direction: column;
