+++ client/src/pages/performance-list/PerformanceList.style.ts
@@ -30,7 +30,7 @@ export const S = {
   `,
   Main: styled.main`
     width: 390px;
-    min-height: calc(100vh - 100px);
+    min-height: calc(100vh - 120px);
     background-color: var(--theme-background-color);
     display: flex;
     flex-direction: column;
